+++ src/services/index.ts
@@ -1,7 +1,9 @@
 import groupService from "./groupService"
 import mainService from "./mainService"
+import userService from "./userService"
 //service index file
 export {
     groupService,
-    mainService
+    mainService,
+    userService
 }
